Remove special fonts
Previously, these dotfiles expected a special font with icon
characters to be installed. This complicated things and wouldn't
work if the user couldn't install fonts.

Dependency on that font has been removed.

This commit removes instructions for installing these these fonts
as well as scripts to help with their installation.