Clean .bashrc colors and style
Previously, the .bashrc had a bit of a mess in its colors and
styles. It used \033 rather than \e. Because colors and styles are
specified using numbers, these extra numbers took a bit more
mental parsing. Additionally, sometimes a color and style was
combined. For example, it might reset style before applying bold.

This commit separates out the colors and styles. It also makes the
colors easier to parse at a glance.